--- scripts/cleanup.py
@@ -1,13 +1,0 @@
-import gc
-import torch
-
-def cleanup_gpu():
-    """
-    Function to clean up GPU memory
-    """
-    # Remove references to all CUDA objects
-    for obj in gc.get_objects():
-        if torch.is_tensor(obj) and obj.is_cuda:
-            del obj
-    gc.collect()
-    torch.cuda.empty_cache()
